Added changes suggest on GitHub. Minor typos/rewording, created test_crop_all file, and updated the raster_stack_crop vignette.
