+++ earthpy/tests/conftest.py
@@ -4,6 +4,8 @@ import numpy as np
 import pytest
 from affine import Affine
 import rasterio as rio
+import geopandas as gpd
+from shapely.geometry import Polygon
 
 
 @pytest.fixture
@@ -157,3 +159,32 @@ def in_paths(basic_image_tif):
 def out_path(tmpdir):
     """ A path for an output .tif file. """
     return os.path.join(str(tmpdir), "out.tif")
+
+
+@pytest.fixture
+def basic_geometry():
+    """
+    A square polygon spanning (2, 2) to (4.25, 4.25) in x and y directions
+    Borrowed from rasterio/tests/conftest.py
+
+    Returns
+    -------
+    dict: GeoJSON-style geometry object.
+        Coordinates are in grid coordinates (Affine.identity()).
+    """
+    return Polygon([(2, 2), (2, 4.25), (4.25, 4.25), (4.25, 2), (2, 2)])
+
+
+@pytest.fixture
+def basic_geometry_gdf(basic_geometry):
+    """
+    A GeoDataFrame containing the basic geometry
+
+    Returns
+    -------
+    GeoDataFrame containing the basic_geometry polygon
+    """
+    gdf = gpd.GeoDataFrame(
+        geometry=[basic_geometry], crs={"init": "epsg:4326"}
+    )
+    return gdf
